Added expanded desktop feature.
This feature gives a cleaner look to the homescreen and app drawer by
only giving access to view the battery status, time and navbar when the
cursor is located near them.
Can be toggled on/off in settings/display.
Only works on the homescreen and app drawer at the moment.