Add option to create folders and rename files.
Use R to create a new folder.
The rename function is a bit buggy. It works sometimes, and doesn't work
sometimes. It brings up the OSK with the file name without any problems,
but when it comes to saviing the file's name it crashes sometimes.

Also .bin files can be viewed using the file manager