Bring up built in dark skin toggle.
This feature can be found in Settings->Developer Options. It is quite
similar to the "Dark Theme" feature found in Android Marshmallow (6.0).

I also decided to get rid of all the dumping stuff, as its rather
useless and just taking up space. We don't need all this functionality
since it is better offered in CFW's and separate homebrews.